Squashed commit of the following:
commit 784267344cab89e9869911c41030143a40ba65c1
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Thu Apr 7 14:45:01 2022 +0300

    Fix check exists already flag

commit f9963b8788aa654f5c04cc90f2959a43147c6f10
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Thu Apr 7 14:29:39 2022 +0300

    typo

commit d4f168a7d63beae1896215a6f03810d3b9d38708
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Thu Apr 7 14:27:15 2022 +0300

    Fix check for flag

commit 60fe2874eac0ae0097eb9c71e07a275aac60f22d
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Thu Apr 7 13:56:54 2022 +0300

    Use GetNodeStateStatsView for getpeerinfo

commit 38cf39d8bd3e02a6cda7adeffaedbc27c8ee4f2f
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Thu Apr 7 13:35:37 2022 +0300

    Only one flag from one to one

commit 9bc358fac55ae4c5a54f47193164fe713486b7db
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Wed Apr 6 16:00:42 2022 +0300

    Added transaction hash to AccountProfile full form

commit 30d2cde02f65e9e52cad63a098761e26d5fe7e18
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Wed Apr 6 15:58:10 2022 +0300

    More fixes

commit 6abb06eb34d594fdf7ce77b7d369b25271b6000e
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Tue Apr 5 18:57:05 2022 +0300

    Times changes

commit aad081a95cf8593dfc3e636d277f8c7bf1cf5344
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Tue Apr 5 16:55:10 2022 +0300

    reducing the impact on the reputation of scores 1,2 for content

commit 1f718a3ea9305709ce5316cd6c41271903c7eeac
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Tue Apr 5 13:29:36 2022 +0300

    Flag ContentHash should be from ContentAddress

commit d37c52cad72a0ff28ea483bcf573079c0cc196f9
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 17:10:28 2022 +0300

    debug test tmp params

commit edaf41a32b3bf3e138760a6fc78858e6b181299d
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 17:05:02 2022 +0300

    Test debug params

commit e6f25be7c4f8b7acd382c48c236effc8d528c49d
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 16:57:40 2022 +0300

    Complete moderation flag model

commit ddd2de6c0d104569397650ba5d45514b838e5a41
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 16:57:28 2022 +0300

    Model for moderation flag

commit 530cfe2f807f59595882628d893fa10da061ee73
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 09:53:37 2022 +0300

    Fix missed comment answers

commit c160b3b9a700652669ebfe4e8221d41586e93f69
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 09:53:21 2022 +0300

    Base class for social transactions

commit 7d1ad7cb9509696297cd12bff177f0b5d3fda793
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 09:52:57 2022 +0300

    Refactoring syntax

commit 87e3f6d16c5551d260b53f48de0c8f38e639f6e9
Author: Andy Oknen <andy.oknen@yandex.ru>
Date:   Mon Apr 4 09:51:17 2022 +0300

    Added System table

+++ src/pocketdb/models/base/Payload.h
@@ -40,8 +40,8 @@ namespace PocketTx
         shared_ptr<string> GetString7() const;
         void SetString7(string value);
 
-        shared_ptr<int> GetInt1() const;
-        void SetInt1(int value);
+        shared_ptr<int64_t> GetInt1() const;
+        void SetInt1(int64_t value);
 
     protected:
 
@@ -53,7 +53,7 @@ namespace PocketTx
         shared_ptr<string> m_string5 = nullptr;
         shared_ptr<string> m_string6 = nullptr;
         shared_ptr<string> m_string7 = nullptr;
-        shared_ptr<int> m_int1 = nullptr;
+        shared_ptr<int64_t> m_int1 = nullptr;
 
     };
